main/requirements.txt
DAWG-Python==0.7.2
Django==2.2.5
djangorestframework==3.10.3
docopt==0.6.2
gunicorn==19.9.0
lxml==4.4.1
psycopg2==2.8.3
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
python-docx==0.8.10
pytz==2019.2
sqlparse==0.3.0
django-cors-headers==3.1.0